% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{beta.ci}
\alias{beta.ci}
\title{Beta Approximation Confidence Intervals}
\usage{
beta.ci(object, conf = 0.95, index = 1)
}
\arguments{
\item{object}{a bootstrap object returned from a call to boot.}

\item{conf}{a scalar or vector containing the confidence level of the 
required interval.}

\item{index}{the index of the statistic of interest within the output of a 
call to boot.out$statistic.}
}
\description{
Using the bootstrapping mean and variance to calculate beta distributed two-sided 
confidence intervals.
}

