% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMixed.R
\name{GetData.Mixed}
\alias{GetData.Mixed}
\title{Get the the long data}
\usage{
GetData.Mixed(object)

GetData.Mixed(object)
}
\arguments{
\item{object}{a Mixed object created by GetMixed}
}
\value{
a dataframe with long format

a dataframe with long format
}
\description{
Get the the long data

Get the the long data
}
