% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMixed.R
\name{GetMixed}
\alias{GetMixed}
\title{Create and return an object of class Mixed}
\usage{
GetMixed(
  dataset,
  validLetters = 2,
  Ids = "ID",
  variableName = "Levels",
  valueNamesP = NULL
)
}
\arguments{
\item{dataset}{A dataset with more than one level}

\item{validLetters}{How many letters in variable names that are significant}

\item{variableName}{Name of the variable having the level number}

\item{valueNamesP}{An optional list with variable names for the variables}
}
\value{
an object of class Mixed to use with lme4
}
\description{
Create and return an object of class Mixed
}
\examples{
obj <- GetMixed(MixedModelData[c("ID","Gender","NeAutT1","NeComT1","HWTeachT1",
"NeComT2","HWTeachT2","NeAutT3","NeComT2","HWTeachT3")], validLetters = 4,
level)
}
