% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedModels-package.R
\docType{package}
\name{MixedModels-package}
\alias{MixedModels-package}
\title{Mixed Models with relationship matrices}
\description{
An R package for (Generalized) Linear Mixed Models (G)LMM which
allows to use user defined 'G-site' covariance relationship matrices.
}
\author{
Jonas Klasen
}

