% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Mixed Models incopareting relationship matrices}
\usage{
mm(formula, data, family = NULL, REML = TRUE, covarrel = list(),
  control = list(), start = NULL, verbose = FALSE, subset, weights,
  na.action, offset, contrasts = NULL, devFunOnly = FALSE, ...)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the
    fixed-effects and random-effects part of the model, with the
    response on the left of a \code{~} operator and the terms, separated
    by \code{+} operators, on the right.  Random-effects terms are
    distinguished by vertical bars (\code{|}) separating expressions
    for design matrices from grouping factors.  Two vertical bars
    (\code{||}) can be used to specify multiple uncorrelated random
    effects for the same grouping variable. (Because of the way it is
    implemented, the \code{||}-syntax \emph{works
    only for design matrices containing
    numeric (continuous) predictors}; to fit models with independent
  categorical effects, see \code{\link{dummy}} or the \code{lmer_alt}
  function from the \code{afex} package.) 
}

\item{data}{an optional data frame containing the variables named in
    \code{formula}.  By default the variables are taken from the
    environment from which \code{lmer} is called. While \code{data} is
    optional, the package authors \emph{strongly} recommend its use,
    especially when later applying methods such as \code{update} and
    \code{drop1} to the fitted model (\emph{such methods are not
    guaranteed to work properly if \code{data} is omitted}). If
    \code{data} is omitted, variables will be taken from the environment
    of \code{formula} (if specified as a formula) or from the parent
    frame (if specified as a character vector).}

\item{family}{a GLM family, see \code{\link[stats]{glm}} and
\code{\link[stats]{family}}.}

\item{REML}{logical scalar - Should the estimates be chosen to
    optimize the REML criterion (as opposed to the log-likelihood)?}

\item{covarrel}{a named list of relationship matrices. The names must
correspond to the names of grouping factors for random-effects terms in the
\code{formula} argument.}

\item{control}{a list (of correct class, resulting from
    \code{\link{lmerControl}()} or \code{\link{glmerControl}()}
    respectively) containing control parameters, including the nonlinear
    optimizer to be used and parameters to be passed through to the
    nonlinear optimizer, see the \code{*lmerControl} documentation for
    details.}

\item{start}{a named \code{\link{list}} of starting values for the
    parameters in the model.  For \code{lmer} this can be a numeric
    vector or a list with one component named \code{"theta"}.}

\item{verbose}{integer scalar.  If \code{> 0} verbose output is
    generated during the optimization of the parameter estimates.  If
    \code{> 1} verbose output is generated during the individual PIRLS
    steps.}

\item{subset}{an optional expression indicating the subset of the rows
    of \code{data} that should be used in the fit. This can be a logical
    vector, or a numeric vector indicating which observation numbers are
    to be included, or a character vector of the row names to be
    included.  All observations are included by default.}

\item{weights}{an optional vector of \sQuote{prior weights} to be used
    in the fitting process.  Should be \code{NULL} or a numeric vector.
    Prior \code{weights} are \emph{not} normalized or standardized in
    any way.  In particular, the diagonal of the residual covariance
    matrix is the squared residual standard deviation parameter
    \code{\link{sigma}} times the vector of inverse \code{weights}.
    Therefore, if the \code{weights} have relatively large magnitudes,
    then in order to compensate, the \code{\link{sigma}} parameter will
    also need to have a relatively large magnitude.}

\item{na.action}{a function that indicates what should happen when the
    data contain \code{NA}s.  The default action (\code{na.omit},
    inherited from the 'factory fresh' value of
    \code{getOption("na.action")}) strips any observations with any
    missing values in any variables.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during
    fitting. This should be \code{NULL} or a numeric vector of length
    equal to the number of cases.  One or more \code{\link{offset}}
    terms can be included in the formula instead or as well, and if more
    than one is specified their sum is used.  See
    \code{\link{model.offset}}.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of
    \code{model.matrix.default}.}

\item{devFunOnly}{logical - return only the deviance evaluation
    function. Note that because the deviance function operates on
    variables stored in its environment, it may not return
    \emph{exactly} the same values on subsequent calls (but the results
    should always be within machine tolerance).}

\item{...}{other potential arguments.  A \code{method} argument was
    used in earlier versions of the package.  Its functionality has been
    replaced by the \code{REML} argument.}
}
\value{
a \code{\linkS4class{pedigreemm}} object.
}
\description{
Fit (Generalized) Linear Mixed Models incopareting user
defined 'G-site' covariance relationship matrices. Derifed from
\code{\link[pedigreemm]{pedigreemm}}.
}
\details{
All arguments to this function are the same as those to the
function \code{\link[lme4]{lmer}} (or in case of family
\code{\link[lme4]{glmer}}) except \code{covarrel} which must be a
named list of relationship matrices.  Each name (frequently
there is only one) must correspond to the name of a grouping factor in a
random-effects term in the \code{formula}.  The observed levels of that
factor must be contained in the column and row names of the relationship
matrix.  For each relationship matrix the (left) Cholesky factor of the
observed levels is calculated and applied to the model matrix for that term.
}
\examples{
# modifide from pedigreemm example
# pedigree and relationship matrix
p1 <- new("pedigree",
          sire = as.integer(c(NA, NA, 1, 1, 4, 5, 1, 4, 4, 5)),
          dam  = as.integer(c(NA, NA, 2, NA, 3, 2, 3, 6, 6, 2)),
          label = as.character(1:10))
A <- pedigreemm::getA(p1)
cholA <- chol(A)
# variance componens
varU <- 2
varE <- 6
# number of observations
rep <- 20
n <- rep * 10
ID <- rep(1:10, each = rep)
# simulated random effects
set.seed(108)
bStar <- rnorm(10)
bStar <- bStar * (sqrt(varU) / sd(bStar[ID]))
b <- as.vector(crossprod(cholA, bStar))
# simulated error
e0 <- rnorm(n)
e0 <- e0 * (sqrt(varE) / sd(e0))
# simulated observations
y <- 10 + b[ID] + e0
# models
fm1 <- pedigreemm(y ~ (1 | ID) , pedigree = list(ID = p1))
fm2 <- mm(formula = y ~ (1 | ID) , covarrel = list(ID = A))

# require(coxme)
# fm3 <- lmekin(y ~ (1 | ID) , varlist = list(ID = A), method = "REML")

}
\references{
Vazquez, A.I., D.M. Bates, G.J.M. Rosa, D. Gianola and K.A.
Weigel. (2010). Technical Note: An R package for fitting generalized linear
mixed models in animal breeding. Journal of Animal Science, 88:497-504.
}
\keyword{models}

