% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{mmBoot}
\alias{mmBoot}
\title{Parametric bootstrapping}
\usage{
mmBoot(model, func, R = 1000, cores = 1)
}
\arguments{
\item{model}{a \code{\link[pedigreemm]{pedigreemm}} model object.}

\item{func}{a function which extracts the parameter of interest from the 
model. This function should have one argument - the model object.}

\item{R}{the number of bootstrapping runs.}

\item{cores}{the number of cores for parallel computing at Unix-like 
machines.}
}
\description{
Parametric bootstrapping for mixed model with relationship matrix
}

