% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relationshipmmPackage.R
\docType{package}
\name{relationshipmm-package}
\alias{relationshipmm}
\alias{relationshipmm-package}
\title{HFit relationship matrices based mixed-effects model}
\description{
Fit relationship matrices based mixed-effects model.
}
\author{
Jonas Klasen
}

