% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rescale}
\alias{rescale}
\title{Rescale a relationship matrix}
\usage{
rescale(x, tolerance = 0.001)
}
\arguments{
\item{x}{a symmetric matrix.}

\item{tolerance}{numeric value giving the tolerance for the rescaling.}
}
\description{
Rescale a relationship matrix for a realistic variance component estimation.
}

