% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMixed.R
\name{save.Mixed}
\alias{save.Mixed}
\title{Save for use in other programs}
\usage{
save.Mixed(obj, path)
}
\arguments{
\item{obj}{A Mixed object created by GetMixed}

\item{path}{A path to where the data should be saved}
}
\value{
an object of type Mixed
}
\description{
Save for use in other programs
}
