% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mm.R
\name{grm}
\alias{grm}
\title{Genetic relationship matrix}
\usage{
grm(x)
}
\arguments{
\item{x}{a matrix (n x p) of n individuals (rows) and p SNPs (columns). The
alleles have to be coded by AA = 0, AB = 0.5, and BB = 1. Missing data are
not yet allowed.}
}
\description{
Computers a genetic relationship matrix (GRM) from SNP data.
}

