% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relationshipmmPackage.R
\docType{package}
\name{relationshipmm-package}
\alias{relationshipmm}
\alias{relationshipmm-package}
\title{Mixed Models with relationship matrices}
\description{
An R package for (Generalized) Linear Mixed Models (G)LMM which
allows to use user defined 'G-site' covariance relationship matrices.
}
\author{
Jonas Klasen
}

